LPS-97123 No need for an interface yet
