--- modules/apps/asset/asset-auto-tagger-opennlp-impl/src/main/java/com/liferay/asset/auto/tagger/opennlp/internal/extractor/external/TextExtractorProviderImpl.java
@@ -1,54 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.asset.auto.tagger.opennlp.internal.extractor.external;
-
-import com.liferay.asset.auto.tagger.opennlp.internal.extractor.TextExtractor;
-import com.liferay.asset.auto.tagger.opennlp.internal.extractor.TextExtractorProvider;
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMap;
-import com.liferay.osgi.service.tracker.collections.map.ServiceTrackerMapFactory;
-
-import org.osgi.framework.BundleContext;
-import org.osgi.service.component.annotations.Activate;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Deactivate;
-
-/**
- * @author Alejandro Tardín
- */
-@Component(service = TextExtractorProvider.class)
-public class TextExtractorProviderImpl implements TextExtractorProvider {
-
-	@Override
-	public TextExtractor getTextExtractor(String className) {
-		return _serviceTrackerMap.getService(className);
-	}
-
-	@Activate
-	protected void activate(BundleContext bundleContext) {
-		_serviceTrackerMap =
-			(ServiceTrackerMap<String, TextExtractor<?>>)
-				(ServiceTrackerMap)ServiceTrackerMapFactory.openSingleValueMap(
-					bundleContext, TextExtractor.class, "model.class.name");
-	}
-
-	@Deactivate
-	protected synchronized void deactivate() {
-		_serviceTrackerMap.close();
-	}
-
-	private volatile ServiceTrackerMap<String, TextExtractor<?>>
-		_serviceTrackerMap;
-
-}
